test: move pacquet installer cache test module
